% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dispersalDates}
\alias{dispersalDates}
\title{(OLD!!!) Function to download and show the segmentation dates for the
different behavioral phases}
\usage{
dispersalDates()
}
\value{
data.frame Data frame containing the segmentation dates
}
\description{
This function is used to download and show the file that contains the
segmentation dates that are used to classify the GPS data into residents vs.
dispersers
}
\examples{
# Check segmentation dates
dispersalDates()
}
