% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dog_download}
\alias{dog_download}
\title{Download GPS data from Dropbox}
\usage{
dog_download(
  x,
  overwrite = F,
  clean = F,
  outdir = getwd(),
  printpath = T,
  legacy = F
)
}
\arguments{
\item{x}{data.frame (or a subset of it) as returned from \code{dog_files()}
containing the files you'd like to download. If you simply want to download
all files, consider using \code{dog_download_all()}.}

\item{overwrite}{logical indicating if you wish to overwrite already existing
files}

\item{clean}{logical indicating if you wish to pre-process the data. If set
to \code{FALSE}, raw data will be stored. If set to \code{TRUE}, the files
will be checked for inconsistencies and they will be corrected (e.g. if there
are differing separators). In addition, collar handling periods will be
removed and dispersal/residency dates will be assigned to each fix. Insteady
of the raw data, you will get a single cleaned dataset. I highly recommend
that you set this option to \code{TRUE}.}

\item{outdir}{character directory where you'd like to store the downloaded
files. By default this is set to your working directory (i.e. to the
directory that is shown by \code{getwd()})}

\item{printpath}{should the path to the final file(s) be returned after the
operations are finished? This is handy if you want to fruther handle the
files after the download.}
}
\value{
character filepath to the downloaded files
}
\description{
This function allows you to download the files you subsetted using
\code{dog_files()}. Files can either be downloaded in their raw format (as
they are on Dropbox) using \code{clean = F}. Alternatively, the files can be
pre-processed using \code{clean = T}. This will: (1) make sure all files are
converted to .csv files with the same separator, (2) special characters such
as '°' are removed, (3) a column is indicating if an individual was
dispersing or resident is added, (4) collar handling periods are removed, (5)
perfect duplicates are removed, (6) unnecessary columns are removed, (6) all
files are finally merged into a single .csv file.
}
\examples{
\dontrun{
# Check all available files on Dropbbox
files <- dog_files(rvc = T)

# Randomly select three of the files
files <- files[sample(nrow(files), 3), ]

# Download them
dog_download(files)
}
}
