% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dog_download_all}
\alias{dog_download_all}
\title{Download all GPS data from Dropbox}
\usage{
dog_download_all(
  rvc = T,
  overwrite = F,
  clean = T,
  outdir = getwd(),
  printpath = T
)
}
\arguments{
\item{rvc}{logical indicating wheter "RVC GPS data" should be searched too.}

\item{overwrite}{logical indicating if you wish to overwrite already existing
files}

\item{clean}{logical indicating if you wish to pre-process the data}

\item{outdir}{character directory where you'd like to store the downloaded
files. By default this is set to your working directory (i.e. \code{getwd()})}

\item{printpath}{should the path to the final file(s) be returned after the
operations are finished?}
}
\value{
character filepath to the downloaded files
}
\description{
This is a wrapper function around \code{dog_download} and allows you to
quickly download all files available on Dropbox.
}
\examples{
\dontrun{
# Download them
dog_download_all(rvc = T)
}
}
