% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dog_files}
\alias{dog_files}
\title{Display all files containing GPS data on Dropbox}
\usage{
dog_files(rvc = F)
}
\arguments{
\item{rvc}{logical indicating wheter "RVC GPS data" should be searched too.}
}
\value{
data.frame Data frame containing information on all files containing
GPS data on Dropbox.
}
\description{
This function is used to retrieve information on all files containing GPS
data in the KML-Dispersers folder on Dropbox. The returned dataframe can be
used to identify files that should be downloaded using \code{dog_download()}.
The first time the function is run, you'll need to allow access to your
dropbox.
}
\examples{
# Check all available files on Dropbbox
dog_files(rvc = T)
}
