% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dog_overview}
\alias{dog_overview}
\title{Get an overview of the downloaded GPS data}
\usage{
dog_overview(x, nrow = NULL, ncol = NULL)
}
\arguments{
\item{x}{data.frame or tibble containing GPS data}

\item{nrow}{numeric value indicating the number of desired rows}

\item{ncol}{numeric value indicating the number of desired columns}
}
\value{
plot of the data
}
\description{
Function to create plots that give an overview of each dog's behavioral
phases based on the GPS data
}
\examples{
\dontrun{
# Download GPS data and create overview plot
data <- dog_download_all(rvc = F)

# Create overview plot
dog_overview(data)
}
}
