% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{.collarDates}
\alias{.collarDates}
\title{Function to download and show the collar dates for each individual}
\usage{
.collarDates()
}
\value{
data.frame Data frame containing the collar dates
}
\description{
This function is used to download and show the file that contains the
dates at which collars were deployed and when they stopped working
}
\examples{
# Check collar dates
.collarDates()
}
