% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{.phaseDates}
\alias{.phaseDates}
\title{Function to download and show the segmentation dates for the different
behavioral phases}
\usage{
.phaseDates()
}
\value{
data.frame Data frame containing the segmentation dates
}
\description{
This function is used to download and show the file that contains the
segmentation dates that are used to classify the GPS data into different
behavioral phases / states
}
\examples{
# Check segmentation dates
.phaseDates()
}
