% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{resampleFixes}
\alias{resampleFixes}
\title{Resample GPS fixes}
\usage{
resampleFixes(data, hours, start, tol = 0.5)
}
\arguments{
\item{data}{Data should be resampled}

\item{hours}{Aspired time between fixes in hours}

\item{start}{Hour at which the resampled tracks should start}

\item{tol}{tolerance (in hours) with regards to the hour at which fixes
should be sampled}
}
\value{
\code{data.frame}
}
\description{
Function to resample fixes to a coarser resolution
}
